作成: WishListDataSource切り分け
